requirements.txt
SpeechRecognition==3.14.3
pyttsx3==2.90
PyAudio==0.2.14
wikipedia==1.4.0
google-generativeai==0.8.4
python-dotenv==1.2.1
opencv-python==4.12.0.88
streamlit==1.52.2
fpdf
pytest